requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg[binary]==3.2.10
alembic==1.16.5
redis==6.4.0
httpx==0.28.1
python-dotenv==1.1.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
reportlab==4.4.3
email-validator==2.3.0
itsdangerous==2.1.2
